feat(878): rename network-toggle.tsx
